Lifted out common, model creation functions.
